chore(storybook): Restructure the internal page template layout modules
Rename config to commonMeta and Layout to PageLayout, matching the public page templates.

Let the Page Layout render the Page component, so the withMenu prop sits next to the menu markup filling the grid area it reserves, instead of being split across two files.
